backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.3
pydantic>=2.9.0
python-dotenv>=1.0.1
langchain>=1.0.0
langchain-core>=1.0.0
langchain-deepseek>=0.1.0
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.3.0
openai>=1.52.0
tiktoken>=0.8.0
requests>=2.32.3
html2text>=2024.2.26
pyyaml>=6.0.2
rank-bm25>=0.2.2
